trbg

+++ shaders/ui/atlas_texture.vert
@@ -9,7 +9,7 @@ layout(location = 1) in vec2 pos;
 layout(location = 2) in vec2 size;
 layout(location = 3) in uint uvStart;
 layout(location = 4) in uint uvSize;
-layout(location = 5) in float z_index;
+layout(location = 5) in int z_index;
 
 layout(location = 0) out vec2 fragUv;
 layout(location = 1) out flat vec4 fragColor;
